Adding "Change content order" page for secondary tags.
