+++ docroot/modules/custom/prisoner_hub_taxonomy_sorting/prisoner_hub_taxonomy_sorting.links.task.yml
@@ -2,3 +2,7 @@ prisoner_hub_taxonomy_sorting.change_episode_order:
   route_name: prisoner_hub_taxonomy_sorting.change_episode_order
   title: 'Change episode order'
   base_route: entity.taxonomy_term.canonical
+prisoner_hub_taxonomy_sorting.change_content_order:
+  route_name: prisoner_hub_taxonomy_sorting.change_content_order
+  title: 'Change content order'
+  base_route: entity.taxonomy_term.canonical
